pamsm: unexport all low-level wrappers
